Sampling planner based on table starter
This new version of sampling planner app implements a *.csv table as a starter to confer default sample types and associated sampling zones values. The table can be backed-up, modified, on a spreadsheet, shared, etc. Also, I intend to modify the mapper app to generate/edit this *.csv file by interacting with a map.